backend/requirements/base.txt
# Django and Core Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
mysqlclient==2.2.0
django-mysql==3.15.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==41.0.7
bcrypt==4.1.2

# API Documentation
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1

# Caching
redis==5.0.1
django-redis==5.4.0

# File Storage
Pillow==10.1.0
boto3==1.34.0
django-storages==1.14.2

# Utilities
python-decouple==3.8
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Monitoring & Logging
sentry-sdk==1.38.0
django-extensions==3.2.3

# Validation & Serialization
marshmallow==3.20.1
django-filter==23.5

# Performance
django-debug-toolbar==4.2.0
django-silk==5.0.4
